django/requirements.txt
gunicorn==20.1.0
amqp==2.6.1
aniso8601==8.1.1
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==4.4.7
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
django-debug-toolbar==3.2.1
django-graphql-jwt==0.3.2
django-redis==5.0.0
et-xmlfile==1.1.0
flower==0.9.7
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.5
graphql-relay==3.1.0
humanize==3.6.0
idna==2.10
jsonschema==3.2.0
kombu==4.6.11
loguru==0.5.3
notifiers==1.2.1
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyrsistent==0.17.3
pytz==2021.1
redis==3.5.3
requests==2.25.1
rfc3987==1.3.8
Rx==1.6.1
selenium==3.141.0
singledispatch==3.6.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
